[#534] Re-apply export test-helper extraction dropped by the rebase
Rebasing onto master (which now carries #1308 worker soft-pause via #1949 and
the #1952 cleanup worker) dropped the merge-commit resolution that had
integrated #1308's untagged worker_pause_test.go with #534's build-tag split:

- worker_pause_test.go (untagged, from #1308) calls newTestFactorySet /
  newTestContext, which #534 had left in the legacy-tagged service_test.go.
  In the default build those symbols are excluded by the tag, so the test no
  longer compiled. Move them to a new UNTAGGED testhelpers_test.go so both the
  default `.inb` build and the legacy_xml_backup build compile against them, and
  drop the now-duplicate definitions (and their unused imports) from
  service_test.go.
- Update worker_pause_test.go's NewExportService call 2->3 args (nil signer) to
  match #534's signature change.

Verified: go build / vet / golangci-lint clean on both tags; full backup +
backupsign + inb + blobkeys + mimekit + bootstrap test sweep green on both.
